+++ test/js/ruby/embed.test.js
@@ -62,4 +62,14 @@ describe("embed", () => {
 
     return expect(content).toMatchFormat();
   });
+
+  test("removes whitespace so embedded parsers don't misinterpret", () => {
+    const content = ruby(`
+      <<~MARKDOWN
+          foo
+      MARKDOWN
+    `);
+
+    return expect(content).toMatchFormat();
+  });
 });