+++ test/js/ruby/nodes/kwargs.test.js
@@ -1,30 +1,8 @@
-const { ruby } = require("../../utils");
-
 describe("kwargs", () => {
-  test("basic", () => {
-    const content = ruby(`
-      def foo(bar: baz)
-      end
-    `);
+  test("basic", () => expect("def foo(bar: baz); end").toMatchFormat());
 
-    return expect(content).toMatchFormat();
-  });
+  test("optional", () => expect("def foo(bar:); end").toMatchFormat());
 
-  test("optional", () => {
-    const content = ruby(`
-      def foo(bar:)
-      end
-    `);
-
-    return expect(content).toMatchFormat();
-  });
-
-  test("double splat", () => {
-    const content = ruby(`
-      def foo(bar:, **baz)
-      end
-    `);
-
-    return expect(content).toMatchFormat();
-  });
+  test("double splat", () =>
+    expect("def foo(bar:, **baz); end").toMatchFormat());
 });